+++ src-tauri/src/ff7/data/mod.rs
@@ -3,9 +3,11 @@ pub mod battle;
 pub mod field;
 pub mod kernel;
 pub mod world;
+pub mod party;
 
 pub use basic::*;
 pub use battle::*;
 pub use field::*;
 pub use kernel::*;
 pub use world::*;
+pub use party::*;