+++ src-tauri/src/ff7/types/basic.rs
@@ -39,6 +39,6 @@ pub struct FF7BasicData {
     pub world_zoom: u16,
     pub world_tilt: u16,
     pub world_speed_multiplier: u8,
-    pub party_members: Vec<u8>,
+    pub party_member_ids: Vec<u8>,
     pub key_items: Vec<u8>,
 }