+++ src-tauri/src/ff7/types/mod.rs
@@ -4,10 +4,12 @@ pub mod field;
 pub mod items;
 pub mod kernel;
 pub mod world;
+pub mod party;
 
 pub use basic::*;
 pub use battle::*;
 pub use field::*;
+pub use party::*;
 pub use world::*;
 // pub use items::*;
 // pub use kernel::*;